\docType{methods}
\name{baseRank}
\alias{baseRank}
\alias{baseRank,RankSummary-method}
\title{Get base ranking}
\arguments{
  \item{x}{a RankSummary object}
}
\value{
  a named vector with the base ranking
}
\description{
  Accessor function to obtain the base ranking from
  RankSummary objects.
}

