\name{method}
\alias{method}
\alias{method,RankSummary-method}
\title{Get method}
\arguments{
  \item{x}{a RankSummary object}
}
\value{
  a string with the name of the base ranking method
}
\description{
  Accessor function to obtain the name of the base ranking
  method from RankSummary objects.
}

