\name{rankCor}
\alias{rankCor}
\alias{rankCor,RankSummary-method}
\title{Get rank correlations}
\arguments{
  \item{x}{a RankSummary object}
}
\value{
  A matrix with Spearman's rank correlation coefficients of
  the three rankings contained in a RankSummary object.
}
\description{
  Accessor function to obtain the rank correlation matrix
  from RankSummary objects.
}

