\docType{data}
\name{salmonellaStability}
\alias{salmonellaStability}
\title{Example stability analysis on a subset of the Salmonella infection
dataset}
\format{a list of RankSummary objects that use different base ranking methods}
\description{
  The package contains the results of an example stability
  analysis for the Salmonella infection dataset.
  RankSummary objects were created for five base ranking
  methods on a subset of the genes.
}
\keyword{datasets}

