\name{show}
\alias{show}
\alias{show,RankSummary-method}
\title{Show function for a RankSummary object.}
\arguments{
  \item{object}{an object of class RankSummary.}
}
\description{
  This prints the number of ranked elements, the base
  ranking method, the top elements of the three rankings
  and the top stable set sizes from a RankSummary object.
}

