\name{stableSetSize}
\alias{stableSetSize}
\alias{stableSetSize,RankSummary-method}
\title{Get stable set sizes}
\arguments{
  \item{x}{a RankSummary object}
}
\value{
  a vector with the per cutoff stable set size.
}
\description{
  Accessor function to obtain the sizes of stable sets per
  cutoff from RankSummary objects.
}

