\docType{methods}
\name{stabRank}
\alias{stabRank}
\alias{stabRank,RankSummary-method}
\title{Get stability ranking}
\arguments{
  \item{x}{a RankSummary object}
}
\value{
  a named vector with the stability ranking
}
\description{
  Accessor function to obtain the stability ranking from
  RankSummary objects.
}

